The one about PyLint. Updated pretty much all python files to conform to PEP8 (mostly). `pylint` is now reporting the following:
Python 2.7
    Your code has been rated at 9.51/10
    1 errors, 13 warnings, 8 conventions, 2 refactors, 0 fatals

Python3.4
    Your code has been rated at 9.58/10
    1 errors, 13 warnings, 4 conventions, 2 refactors, 0 fatals

Wildcard and unused imports need to be cleaned up.
Unused variables in vim API functions can be obfuscated by renaming them to `_`.  I'm not sure that's worth the loss of readability.
Python2.7 is complaining about parens around print statements.  We should probably start using `logging` anyway and just tying the logger to stdout when we want to print.
The single error is `Unable to import 'vim'`, which is normal outside of the vim environment.  Probably not worth mocking it just for PyLint.